application-manager/requirements.txt
#flask-appconfig>=0.10
flask-bootstrap
flask-nav
flask-wtf
flask-oauthlib
requests
python-jose